+++ pkg/healthcheck/socket_control_other.go
@@ -1,0 +1,9 @@
+//go:build !linux
+
+package healthcheck
+
+import "syscall"
+
+func bindToDevice(conn syscall.RawConn, ifname string) error {
+	return nil
+}